+++ src/pkgs/controller/__init__.py
@@ -8,7 +8,7 @@ class Controller:
     """
     Class implementing the controller function.
     """
-    CONFIG_ROOT_DIR = './src/controller/configs/'
+    CONFIG_ROOT_DIR = './src/pkgs/controller/configs/'
     CTRL_FRAME_RATE = 10
 
     def __init__(self, idx, name, app, ndigit=2):
